celery-task-queue-example/requirements.txt
amqp
billiard
blinker
celery
click
click-didyoumean
click-plugins
click-repl
Flask
itsdangerous
Jinja2
kombu
MarkupSafe
prompt-toolkit
python-dateutil
python-dotenv
redis
six
tzdata
vine
wcwidth
Werkzeug
